services/ai-service/pyproject.toml
[tool.poetry]
name = "congent"
version = "0.1.0"
description = "agentic content generation from long-form transcripts"
authors = ["Jorge Lewis <31695089+antopiahk@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.112.2"
langchain = "0.2.15"
langgraph-checkpoint = "1.0.8"
langchain-openai = "0.1.23"
langchain-core = "0.2.37"
langchain-community = "0.2.15"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"
langgraph = "^0.2.16"
supabase = "^2.7.4"
pandas = "^2.2.2"
graphrag = "0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
